overrides creator init
