+++ administrator/language/en-GB/com_templates.ini
@@ -261,3 +261,4 @@ COM_TEMPLATES_LEGACY="Legacy"
 COM_TEMPLATES_UNINSTALL="Uninstall template"
 COM_TEMPLATES_FORK="Duplicate template"
 COM_TEMPLATE_EDIT_FILES="Edit template files"
+COM_TEMPLATES_CREATE_OVERRIDE="Create Override"
